// frontend/src/constants/types.ts
// User type definition
export interface User {
  f_id: string;
  f_username: string;
  f_password?: string;
}

// Employee type definition
export interface Employee {
  f_Id: string;
  f_image?: string;
  f_name: string;
  f_email: string;
  f_mobile_no: string;
  f_designation: string;
  f_gender: string;
  f_course: string;
  f_created_date: string;
}